# docs/swagger.yaml
basePath: /
definitions:
  swagger.Login:
    properties:
      access_key:
        type: string
      secret_key:
        type: string
    type: object
host: localhost:8082
info:
  contact: {}
  description: This is a Minio Server
  title: Swagger Example API
  version: 0.0.1
paths:
  /api/buckets/exists:
    get:
      consumes:
      - application/json
      responses:
        "200":
          description: '{"message": "获取成功", status: 1}'
        "400":
          description: '{"message": "获取失败", status: -1}'
      security:
      - ApiKeyAuth: []
      - ApiKeyXRole: []
      summary: 获取存储桶详细信息
      tags:
      - buckets
  /api/buckets/list:
    get:
      consumes:
      - application/json
      responses:
        "200":
          description: '{"message": "获取成功", status: 1}'
        "400":
          description: '{"message": "获取失败", status: -1}'
      security:
      - ApiKeyAuth: []
      summary: 获取存储桶列表
      tags:
      - buckets
  /api/buckets/listobjects:
    get:
      consumes:
      - application/json
      parameters:
      - description: 存储桶名必传
        in: query
        name: bucket
        required: true
        type: string
      responses:
        "200":
          description: '{"message": "获取成功", status: 1}'
        "400":
          description: '{"message": "获取失败", status: -1}'
      security:
      - ApiKeyAuth: []
      summary: 获取存储桶内所有文件列表
      tags:
      - buckets
  /api/buckets/remove:
    post:
      consumes:
      - application/json
      parameters:
      - description: 存储桶名必传
        in: query
        name: bucket
        required: true
        type: string
      responses:
        "200":
          description: '{"message": "删除成功", status: 1}'
        "400":
          description: '{"message": "删除失败", status: -1}'
      security:
      - ApiKeyAuth: []
      - ApiKeyXRole: []
      summary: 删除存储桶
      tags:
      - buckets
  /api/object/remove:
    post:
      consumes:
      - application/json
      parameters:
      - description: 存储桶名必传
        in: query
        name: bucket
        required: true
        type: string
      - description: 文件名必传
        in: query
        name: object
        required: true
        type: string
      responses:
        "200":
          description: '{"message": "删除成功", status: 1}'
        "400":
          description: '{"message": "删除失败", status: -1}'
      security:
      - ApiKeyAuth: []
      - ApiKeyXRole: []
      summary: 删除文件
      tags:
      - objects
  /api/object/upload:
    post:
      consumes:
      - application/json
      - multipart/form-data
      parameters:
      - description: 文件名必传
        in: formData
        name: file
        required: true
        type: file
      - description: 存储桶名必传
        in: formData
        name: bucket
        required: true
        type: string
      responses:
        "200":
          description: '{"message": "删除成功", status: 1}'
        "400":
          description: '{"message": "删除失败", status: -1}'
      security:
      - ApiKeyAuth: []
      - ApiKeyXRole: []
      summary: 上传文件
      tags:
      - objects
  /api/object/url:
    get:
      consumes:
      - application/json
      parameters:
      - description: 存储桶名必传
        in: query
        name: bucket
        required: true
        type: string
      - description: 文件名必传
        in: query
        name: object
        required: true
        type: string
      responses:
        "200":
          description: '{"message": "获取成功", status: 1}'
        "400":
          description: '{"message": "获取失败", status: -1}'
      security:
      - ApiKeyAuth: []
      - ApiKeyXRole: []
      summary: 获取文件的url
      tags:
      - objects
  /api/user/info:
    get:
      consumes:
      - application/json
      description: '注意: header设置token时前面带 Bearer + 空格'
      responses:
        "200":
          description: '{"message": "获取成功", status: 1}'
        "400":
          description: '{"message": "获取失败", status: -1}'
      security:
      - ApiKeyAuth: []
      summary: 获取用户信息
      tags:
      - users
  /api/user/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 账号密码必须填
        in: body
        name: bucket
        required: true
        schema:
          $ref: '#/definitions/swagger.Login'
      responses:
        "200":
          description: '{"message": "登录成功", status: 1}'
        "400":
          description: '{"message": "登录失败", status: -1}'
      summary: 登录
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  ApiKeyXRole:
    in: header
    name: x-role
    type: apiKey
swagger: "2.0"